feat: 🎸  run as a cli command instead of using cloudformation
Cloudformation was problematic as it didn't treat the rate limit with
the SSM Api well. It was originally built as a quick win solution,
knowing it would hammer the SSM Api eventually.   That day came and now
we have a CLI solution.

Note that this solution is likely temporary (more temporary than the
previous method?) as it currently is designed to run in a codebuild
step, but triggering a lambda directly or a step function workflow would
be much faster.