feat: 🎸  move ssm parameter updates into a more intelligent lambda
Instead of using the cloudformation SSM Parameter type, which was
causing load on the PutParameter API due to parallel stacks and no rate
limiting management, this PR moves the updates into a custom resource.

The custom resource leverages the v3 sdk AdaptiveRateLimiter, and to
reduce unecessary calls to the API, an idempotency check is first made,
by a GetParametersByPath call at each hierarchy level.  Parameters that
do not require updates are just skipped in the update process.

This means for a single change to a single configuration key, one
PutParameter call will happen.

The AdaptiveRateLimiter also enables a large batch of initial parameters
on the first deploy, as it is resilient to ThrottlingExceptions.

Rate limiting, and the backoff of API calls due to ThrottlingExceptions
can still make this fail however, due to the configurable lambda
timeout.  There is a sane default set, however if you wish to override
that, you may.

You may also override an initial Max Sleep duration, which in the event
of multiple stacks, spreads the execution out, to reduce the
ThrottlingExceptions, and therefore reduce the backoff time in each SDK
API Call.